Add permutation indexing atoms
Index into lexicographically sorted list of permutations, or find indexes of permutations without constructing the sorted lists.
Worthy of note is the slightly odd monad, Œ?, which will find the longest possible permutation of items [1,length] that would have such an index, but is useful since that range is (length-1)!+1 to length! inclusive. If one needs to find lower entries use the dyad with a range instead.